# Improve PartialEq for slices