@alexcrichton 2015-07-08
@bors: r+ 9f4d5b4be10ce9683569a0bd55f555d30f70b5a5

Nice wins!