@lilianmoraru 2015-07-14
That's interesting. Now I am just a little bit confused.
I thought that using iterators is faster in Rust than using for loops because of bounds checking.
Long time ago I saw a benchmark style test that was initially using a for loop and had pretty bad performance compared to C and C++, then somebody else switches it to iterators and the performance was comparable...